Cleanup account data when closing the last mailbox
Imagine the sequence unmailboxes *; change-folder. The currently open
mailbox cannot be removed by unmailboxes *, which only marks it as
hidden. Then, on change-folder, the current mailbox is closed, but
nobody notifies the Account that the mailbox is gone, so in case of IMAP
the connection to the old server stays around.